refactor: move archiveformat code inside archive pipe
It was being used only in that pipe anyway.
